+++ src/main.tsx
@@ -1,17 +1,33 @@
 import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
-import { getCachedFilters } from './filters-cache.ts';
+import { keepOnlyValidRepositories, getCachedFilters } from './filters-cache.ts';
 import { fetchLockfiles } from './fetch-lockfiles.ts';
 import './index.css';
 import { ThemeType, DefiniteThemeType, getThemePreference, applyTheme } from './theme.tsx';
+import { sha256 } from './utils.ts';
+
+function getDefaultSelectedRepos(repoHashes: Record<string, string>): Set<string> {
+  const filteredRepos = new URLSearchParams(window.location.search).get('repos');
+  if (filteredRepos) {
+    const desiredRepos = filteredRepos
+      .split(',')
+      .map((hash) => repoHashes[hash])
+      .filter((x) => x !== undefined);
+    return new Set(desiredRepos);
+  } else {
+    return getCachedFilters();
+  }
+}
 
 // eslint-disable-next-line react-refresh/only-export-components
 function Root({ lockfilesUrl }: { lockfilesUrl: string }) {
   const [repositories, setRepositories] = useState<null | string[]>(null);
+  const [repoHashes, setRepoHashes] = useState<null | Record<string, string>>(null);
   const [lockfiles, setLockfiles] = useState<null | LockfilesMap>(null);
   const [systemTheme, setSystemTheme] = useState<null | DefiniteThemeType>(null);
   const [theme, setTheme] = useState<null | ThemeType>(null);
+  const [defaultSelectedRepos, setDefaultSelectedRepos] = useState<null | Set<string>>(null);
 
   useEffect(() => {
     let mounted = true;
@@ -60,12 +76,47 @@ function Root({ lockfilesUrl }: { lockfilesUrl: string }) {
     };
   }, []);
 
-  return lockfiles !== null && repositories !== null && systemTheme !== null && theme !== null ? (
+  useEffect(() => {
+    if (repositories !== null && repoHashes !== null) {
+      const repositoriesSubset = getDefaultSelectedRepos(repoHashes);
+      const repositoriesAsSet = new Set(repositories);
+      // Remove any previously selected repository that is no longer available
+      const cleanedSelectedRepos = keepOnlyValidRepositories(repositoriesSubset, repositoriesAsSet);
+      const newSelectedRepos = cleanedSelectedRepos.size > 0 ? cleanedSelectedRepos : repositoriesAsSet;
+      setDefaultSelectedRepos(newSelectedRepos);
+    }
+  }, [repositories, repoHashes]);
+
+  useEffect(() => {
+    (async () => {
+      if (repositories !== null) {
+        const fullHashes = await Promise.all(repositories.map((repo) => sha256(repo)));
+        const shortHashes = fullHashes.map((hash) => hash.slice(0, 7));
+        setRepoHashes(
+          shortHashes.reduce(
+            (acc, hash, i) => {
+              acc[hash] = repositories[i];
+              return acc;
+            },
+            {} as Record<string, string>
+          )
+        );
+      }
+    })();
+  }, [repositories]);
+
+  return lockfiles !== null &&
+    repositories !== null &&
+    repoHashes !== null &&
+    systemTheme !== null &&
+    theme !== null &&
+    defaultSelectedRepos !== null ? (
     <App
       repositories={repositories}
+      repoHashes={repoHashes}
       lockfiles={lockfiles}
       baseRepoUrl={window.baseRepoUrl}
-      defaultSelectedRepos={getCachedFilters()}
+      defaultSelectedRepos={defaultSelectedRepos}
       defaultQuery={new URLSearchParams(window.location.search).get('q') || ''}
       systemTheme={systemTheme}
       defaultTheme={theme === 'auto' ? systemTheme : theme}